# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Urja Meter Ops API",
    "description": "Clean API over the Urja Meter Ops portal.",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/meters": {
      "get": {
        "tags": [
          "meters"
        ],
        "summary": "List Meters",
        "operationId": "list_meters_meters_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "matches meter number or serial number",
              "default": "",
              "title": "Search"
            },
            "description": "matches meter number or serial number"
          },
          {
            "name": "meter_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Meter Id"
            }
          },
          {
            "name": "serial_no",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Serial No"
            }
          },
          {
            "name": "make",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Make"
            }
          },
          {
            "name": "phase_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Phase Type"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "dt_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dt Code"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeterList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/meters/{meter_id}": {
      "get": {
        "tags": [
          "meters"
        ],
        "summary": "Get Meter Detail",
        "operationId": "get_meter_detail_meters__meter_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "meter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Meter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeterDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/network/dts": {
      "get": {
        "tags": [
          "network"
        ],
        "summary": "List Dts",
        "operationId": "list_dts_network_dts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "matches DT code, name, or feeder code",
              "default": "",
              "title": "Search"
            },
            "description": "matches DT code, name, or feeder code"
          },
          {
            "name": "dt_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dt Code"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            }
          },
          {
            "name": "feeder_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Feeder Code"
            }
          },
          {
            "name": "capacity_kva",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Capacity Kva"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DtList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/meters/{meter_id}/consumption": {
      "get": {
        "tags": [
          "consumption"
        ],
        "summary": "Get Meter Consumption",
        "operationId": "get_meter_consumption_meters__meter_id__consumption_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "meter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Meter Id"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "only readings at or after this timestamp",
              "title": "Start"
            },
            "description": "only readings at or after this timestamp"
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "only readings at or before this timestamp",
              "title": "End"
            },
            "description": "only readings at or before this timestamp"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 1,
              "default": 100,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnergyReadingList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/cache/clear": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Clear Cache",
        "operationId": "clear_cache_admin_cache_clear_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Clear Cache Admin Cache Clear Post"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "Dt": {
        "properties": {
          "dt_code": {
            "type": "string",
            "title": "Dt Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "feeder_code": {
            "type": "string",
            "title": "Feeder Code"
          },
          "capacity_kva": {
            "type": "integer",
            "title": "Capacity Kva"
          }
        },
        "type": "object",
        "required": [
          "dt_code",
          "name",
          "feeder_code",
          "capacity_kva"
        ],
        "title": "Dt"
      },
      "DtList": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Dt"
            },
            "type": "array",
            "title": "Data"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "data",
          "total",
          "page",
          "page_size"
        ],
        "title": "DtList"
      },
      "EnergyReading": {
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "kwh": {
            "type": "number",
            "title": "Kwh"
          },
          "kvah": {
            "type": "number",
            "title": "Kvah"
          },
          "volt_r": {
            "type": "number",
            "title": "Volt R"
          }
        },
        "type": "object",
        "required": [
          "timestamp",
          "kwh",
          "kvah",
          "volt_r"
        ],
        "title": "EnergyReading"
      },
      "EnergyReadingList": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/EnergyReading"
            },
            "type": "array",
            "title": "Data"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "data",
          "total",
          "page",
          "page_size"
        ],
        "title": "EnergyReadingList"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HierarchyLevel": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "code": {
            "type": "string",
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "name",
          "code"
        ],
        "title": "HierarchyLevel"
      },
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest"
      },
      "Meter": {
        "properties": {
          "meter_id": {
            "type": "string",
            "title": "Meter Id"
          },
          "serial_no": {
            "type": "string",
            "title": "Serial No"
          },
          "make": {
            "type": "string",
            "title": "Make"
          },
          "phase_type": {
            "type": "string",
            "title": "Phase Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "dt_code": {
            "type": "string",
            "title": "Dt Code"
          }
        },
        "type": "object",
        "required": [
          "meter_id",
          "serial_no",
          "make",
          "phase_type",
          "status",
          "dt_code"
        ],
        "title": "Meter"
      },
      "MeterDetail": {
        "properties": {
          "meter_id": {
            "type": "string",
            "title": "Meter Id"
          },
          "serial_no": {
            "type": "string",
            "title": "Serial No"
          },
          "make": {
            "type": "string",
            "title": "Make"
          },
          "phase_type": {
            "type": "string",
            "title": "Phase Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "installation_type": {
            "type": "string",
            "title": "Installation Type"
          },
          "zone": {
            "$ref": "#/components/schemas/HierarchyLevel"
          },
          "circle": {
            "$ref": "#/components/schemas/HierarchyLevel"
          },
          "division": {
            "$ref": "#/components/schemas/HierarchyLevel"
          },
          "subdivision": {
            "$ref": "#/components/schemas/HierarchyLevel"
          },
          "substation": {
            "$ref": "#/components/schemas/HierarchyLevel"
          },
          "feeder": {
            "$ref": "#/components/schemas/HierarchyLevel"
          },
          "dt": {
            "$ref": "#/components/schemas/HierarchyLevel"
          }
        },
        "type": "object",
        "required": [
          "meter_id",
          "serial_no",
          "make",
          "phase_type",
          "status",
          "installation_type",
          "zone",
          "circle",
          "division",
          "subdivision",
          "substation",
          "feeder",
          "dt"
        ],
        "title": "MeterDetail"
      },
      "MeterList": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Meter"
            },
            "type": "array",
            "title": "Data"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "data",
          "total",
          "page",
          "page_size"
        ],
        "title": "MeterList"
      },
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "TokenResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}